pyproject.toml
[tool.poetry]
name = "nfl-screenshot"
version = "0.1.0"
description = "NFL stats page screenshot tool"
authors = ["Ernesto Martinez"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,!=3.9.7"
playwright = "^1.55.0"
anthropic = "^0.40.0"
python-dotenv = "^1.0.0"
tqdm = "^4.67.0"
inquirer = "^3.4.0"
ratelimit = "^2.2.1"
rich = "^14.2.0"
streamlit = "^1.40.0"
plotly = "^5.24.0"
beautifulsoup4 = "^4.12.0"
pyyaml = "^6.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.1"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
